requirements.txt
livekit-agents>=0.7.0
livekit-plugins-deepgram>=0.2.0
livekit-plugins-openai>=0.2.0
httpx>=0.27.0
pydantic>=2.8.0
python-dotenv>=1.0.1
loguru>=0.7.2
uvloop>=0.19.0